+++ rivertest/rivertest_test.go
@@ -40,10 +40,11 @@ func TestRequireInserted(t *testing.T) {
 	ctx := context.Background()
 
 	type testBundle struct {
-		dbPool *pgxpool.Pool
-		driver *riverpgxv5.Driver
-		mockT  *testutil.MockT
-		schema string
+		dbPool     *pgxpool.Pool
+		driver     *riverpgxv5.Driver
+		mockT      *testutil.MockT
+		schema     string
+		schemaOpts *RequireInsertedOpts
 	}
 
 	setup := func(t *testing.T) (*river.Client[pgx.Tx], *testBundle) {
@@ -61,10 +62,11 @@ func TestRequireInserted(t *testing.T) {
 		require.NoError(t, err)
 
 		return riverClient, &testBundle{
-			dbPool: dbPool,
-			driver: driver,
-			mockT:  testutil.NewMockT(t),
-			schema: schema,
+			dbPool:     dbPool,
+			driver:     driver,
+			mockT:      testutil.NewMockT(t),
+			schema:     schema,
+			schemaOpts: &RequireInsertedOpts{Schema: schema},
 		}
 	}
 
@@ -76,7 +78,7 @@ func TestRequireInserted(t *testing.T) {
 		_, err := riverClient.Insert(ctx, Job1Args{String: "foo"}, nil)
 		require.NoError(t, err)
 
-		job := requireInserted(ctx, t, bundle.driver, bundle.schema, &Job1Args{}, nil)
+		job := requireInserted(ctx, t, bundle.driver, &Job1Args{}, bundle.schemaOpts)
 		require.False(t, bundle.mockT.Failed)
 		require.Equal(t, "foo", job.Args.String)
 	})
@@ -112,7 +114,7 @@ func TestRequireInsertedTx(t *testing.T) {
 		_, err := riverClient.InsertTx(ctx, bundle.tx, Job1Args{String: "foo"}, nil)
 		require.NoError(t, err)
 
-		job := requireInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, emptySchema, &Job1Args{}, nil)
+		job := requireInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, &Job1Args{}, nil)
 		require.False(t, bundle.mockT.Failed)
 		require.Equal(t, "foo", job.Args.String)
 	})
@@ -128,11 +130,11 @@ func TestRequireInsertedTx(t *testing.T) {
 		_, err = riverClient.InsertTx(ctx, bundle.tx, Job2Args{Int: 123}, nil)
 		require.NoError(t, err)
 
-		job1 := requireInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, emptySchema, &Job1Args{}, nil)
+		job1 := requireInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, &Job1Args{}, nil)
 		require.False(t, bundle.mockT.Failed)
 		require.Equal(t, "foo", job1.Args.String)
 
-		job2 := requireInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, emptySchema, &Job2Args{}, nil)
+		job2 := requireInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, &Job2Args{}, nil)
 		require.False(t, bundle.mockT.Failed)
 		require.Equal(t, 123, job2.Args.Int)
 	})
@@ -149,12 +151,12 @@ func TestRequireInsertedTx(t *testing.T) {
 		require.NoError(t, err)
 
 		// Visible in the original transaction.
-		job := requireInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, &Job1Args{}, nil)
+		job := requireInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, &Job1Args{}, nil)
 		require.False(t, bundle.mockT.Failed)
 		require.Equal(t, "foo", job.Args.String)
 
 		// Not visible in the second transaction.
-		_ = requireInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, otherTx, emptySchema, &Job1Args{}, nil)
+		_ = requireInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, otherTx, &Job1Args{}, nil)
 		require.True(t, bundle.mockT.Failed)
 	})
 
@@ -166,7 +168,7 @@ func TestRequireInsertedTx(t *testing.T) {
 		_, err := riverClient.InsertTx(ctx, bundle.tx, Job1Args{String: "foo"}, nil)
 		require.NoError(t, err)
 
-		_ = requireInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, &Job2Args{}, nil)
+		_ = requireInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, &Job2Args{}, nil)
 		require.True(t, bundle.mockT.Failed)
 		require.Equal(t,
 			failureString("No jobs found with kind: job2")+"\n",
@@ -184,7 +186,7 @@ func TestRequireInsertedTx(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		_ = requireInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, &Job1Args{}, nil)
+		_ = requireInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, &Job1Args{}, nil)
 		require.True(t, bundle.mockT.Failed)
 		require.Equal(t,
 			failureString("More than one job found with kind: job1 (you might want RequireManyInserted instead)")+"\n",
@@ -199,7 +201,7 @@ func TestRequireInsertedTx(t *testing.T) {
 		_, err := riverClient.InsertTx(ctx, bundle.tx, Job2Args{Int: 123}, nil)
 		require.NoError(t, err)
 
-		_ = requireInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, &Job1Args{}, nil)
+		_ = requireInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, &Job1Args{}, nil)
 		require.True(t, bundle.mockT.Failed)
 		require.Equal(t,
 			failureString("No jobs found with kind: job1")+"\n",
@@ -239,7 +241,7 @@ func TestRequireInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := sameOpts()
 			opts.MaxAttempts = 77
-			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' max attempts 78 not equal to expected 77")+"\n",
@@ -250,7 +252,7 @@ func TestRequireInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := sameOpts()
 			opts.Priority = 3
-			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' priority 2 not equal to expected 3")+"\n",
@@ -261,7 +263,7 @@ func TestRequireInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := sameOpts()
 			opts.Queue = "wrong_queue"
-			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' queue 'another_queue' not equal to expected 'wrong_queue'")+"\n",
@@ -272,7 +274,7 @@ func TestRequireInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := sameOpts()
 			opts.ScheduledAt = testTime.Add(3*time.Minute + 23*time.Second + 123*time.Microsecond)
-			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' scheduled at 2023-10-30T10:45:23.000123Z not equal to expected 2023-10-30T10:48:46.000246Z")+"\n",
@@ -283,7 +285,7 @@ func TestRequireInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := sameOpts()
 			opts.State = rivertype.JobStateCancelled
-			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' state 'scheduled' not equal to expected 'cancelled'")+"\n",
@@ -294,7 +296,7 @@ func TestRequireInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := sameOpts()
 			opts.Tags = []string{"tag2"}
-			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' tags [tag1] not equal to expected [tag2]")+"\n",
@@ -306,7 +308,7 @@ func TestRequireInsertedTx(t *testing.T) {
 			opts := emptyOpts()
 			opts.MaxAttempts = job.MaxAttempts
 			opts.Priority = job.Priority
-			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.False(t, mockT.Failed, "Should have succeeded, but failed with: "+mockT.LogOutput())
 		})
 
@@ -315,7 +317,7 @@ func TestRequireInsertedTx(t *testing.T) {
 			opts := sameOpts()
 			opts.MaxAttempts = 77
 			opts.Priority = 3
-			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			_ = requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' max attempts 78 not equal to expected 77, priority 2 not equal to expected 3")+"\n",
@@ -325,7 +327,7 @@ func TestRequireInsertedTx(t *testing.T) {
 		t.Run("AllSameSucceeds", func(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := sameOpts()
-			requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			requireInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.False(t, mockT.Failed)
 		})
 	})
@@ -339,10 +341,11 @@ func TestRequireNotInserted(t *testing.T) {
 	ctx := context.Background()
 
 	type testBundle struct {
-		dbPool *pgxpool.Pool
-		driver *riverpgxv5.Driver
-		mockT  *testutil.MockT
-		schema string
+		dbPool     *pgxpool.Pool
+		driver     *riverpgxv5.Driver
+		mockT      *testutil.MockT
+		schema     string
+		schemaOpts *RequireInsertedOpts
 	}
 
 	setup := func(t *testing.T) (*river.Client[pgx.Tx], *testBundle) {
@@ -360,10 +363,11 @@ func TestRequireNotInserted(t *testing.T) {
 		require.NoError(t, err)
 
 		return riverClient, &testBundle{
-			dbPool: dbPool,
-			driver: driver,
-			mockT:  testutil.NewMockT(t),
-			schema: schema,
+			dbPool:     dbPool,
+			driver:     driver,
+			mockT:      testutil.NewMockT(t),
+			schema:     schema,
+			schemaOpts: &RequireInsertedOpts{Schema: schema},
 		}
 	}
 
@@ -375,7 +379,7 @@ func TestRequireNotInserted(t *testing.T) {
 		_, err := riverClient.Insert(ctx, Job2Args{Int: 123}, nil)
 		require.NoError(t, err)
 
-		requireNotInserted(ctx, t, bundle.driver, bundle.schema, &Job1Args{}, nil)
+		requireNotInserted(ctx, t, bundle.driver, &Job1Args{}, bundle.schemaOpts)
 		require.False(t, bundle.mockT.Failed)
 	})
 }
@@ -410,7 +414,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 		_, err := riverClient.InsertTx(ctx, bundle.tx, Job2Args{Int: 123}, nil)
 		require.NoError(t, err)
 
-		requireNotInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, emptySchema, &Job1Args{}, nil)
+		requireNotInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, &Job1Args{}, nil)
 		require.False(t, bundle.mockT.Failed)
 	})
 
@@ -419,10 +423,10 @@ func TestRequireNotInsertedTx(t *testing.T) {
 
 		_, bundle := setup(t)
 
-		requireNotInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, emptySchema, &Job1Args{}, nil)
+		requireNotInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, &Job1Args{}, nil)
 		require.False(t, bundle.mockT.Failed)
 
-		requireNotInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, emptySchema, &Job2Args{}, nil)
+		requireNotInsertedTx[*riverpgxv5.Driver](ctx, t, bundle.tx, &Job2Args{}, nil)
 		require.False(t, bundle.mockT.Failed)
 	})
 
@@ -438,11 +442,11 @@ func TestRequireNotInsertedTx(t *testing.T) {
 		require.NoError(t, err)
 
 		// Not visible in the second transaction.
-		requireNotInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, otherTx, emptySchema, &Job1Args{}, nil)
+		requireNotInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, otherTx, &Job1Args{}, nil)
 		require.False(t, bundle.mockT.Failed)
 
 		// Visible in the original transaction.
-		requireNotInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, &Job1Args{}, nil)
+		requireNotInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, &Job1Args{}, nil)
 		require.True(t, bundle.mockT.Failed)
 	})
 
@@ -451,7 +455,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 
 		_, bundle := setup(t)
 
-		requireNotInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, &Job2Args{}, nil)
+		requireNotInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, &Job2Args{}, nil)
 		require.False(t, bundle.mockT.Failed)
 	})
 
@@ -466,7 +470,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		requireNotInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, &Job1Args{}, nil)
+		requireNotInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, &Job1Args{}, nil)
 		require.True(t, bundle.mockT.Failed)
 		require.Equal(t,
 			failureString("2 jobs found with kind, but expected to find none: job1")+"\n",
@@ -481,7 +485,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 		_, err := riverClient.InsertTx(ctx, bundle.tx, Job2Args{Int: 123}, nil)
 		require.NoError(t, err)
 
-		requireNotInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, &Job1Args{}, nil)
+		requireNotInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, &Job1Args{}, nil)
 		require.False(t, bundle.mockT.Failed)
 	})
 
@@ -518,7 +522,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := emptyOpts()
 			opts.MaxAttempts = job.MaxAttempts
-			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' max attempts equal to excluded %d", job.MaxAttempts)+"\n",
@@ -529,7 +533,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := emptyOpts()
 			opts.Priority = job.Priority
-			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' priority equal to excluded %d", job.Priority)+"\n",
@@ -540,7 +544,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := emptyOpts()
 			opts.Queue = job.Queue
-			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' queue equal to excluded '%s'", job.Queue)+"\n",
@@ -551,7 +555,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := emptyOpts()
 			opts.ScheduledAt = job.ScheduledAt
-			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' scheduled at equal to excluded %s", opts.ScheduledAt.Format(rfc3339Micro))+"\n",
@@ -562,7 +566,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := emptyOpts()
 			opts.State = job.State
-			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' state equal to excluded '%s'", job.State)+"\n",
@@ -573,7 +577,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := emptyOpts()
 			opts.Tags = job.Tags
-			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' tags equal to excluded %+v", job.Tags)+"\n",
@@ -585,7 +589,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 			opts := emptyOpts()
 			opts.MaxAttempts = job.MaxAttempts // one property matches job, but the other does not
 			opts.Priority = 3
-			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.False(t, mockT.Failed, "Should have succeeded, but failed with: "+mockT.LogOutput())
 		})
 
@@ -594,7 +598,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 			opts := emptyOpts()
 			opts.MaxAttempts = job.MaxAttempts
 			opts.Priority = job.Priority
-			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' max attempts equal to excluded %d, priority equal to excluded %d", job.MaxAttempts, job.Priority)+"\n",
@@ -604,7 +608,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 		t.Run("AllSameFails", func(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := sameOpts()
-			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' max attempts equal to excluded %d, priority equal to excluded %d, queue equal to excluded '%s', scheduled at equal to excluded %s, state equal to excluded '%s', tags equal to excluded %+v", job.MaxAttempts, job.Priority, job.Queue, job.ScheduledAt.Format(rfc3339Micro), job.State, job.Tags)+"\n",
@@ -620,7 +624,7 @@ func TestRequireNotInsertedTx(t *testing.T) {
 			mockT := testutil.NewMockT(t)
 			opts := emptyOpts()
 			opts.Priority = 3
-			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, &Job2Args{}, opts)
+			requireNotInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, &Job2Args{}, opts)
 			require.True(t, mockT.Failed)
 			require.Equal(t,
 				failureString("Job with kind 'job2' priority equal to excluded %d", 3)+"\n",
@@ -637,10 +641,11 @@ func TestRequireManyInserted(t *testing.T) {
 	ctx := context.Background()
 
 	type testBundle struct {
-		dbPool *pgxpool.Pool
-		driver *riverpgxv5.Driver
-		mockT  *testutil.MockT
-		schema string
+		dbPool     *pgxpool.Pool
+		driver     *riverpgxv5.Driver
+		mockT      *testutil.MockT
+		schema     string
+		schemaOpts *RequireInsertedOpts
 	}
 
 	setup := func(t *testing.T) (*river.Client[pgx.Tx], *testBundle) {
@@ -658,10 +663,11 @@ func TestRequireManyInserted(t *testing.T) {
 		require.NoError(t, err)
 
 		return riverClient, &testBundle{
-			dbPool: dbPool,
-			driver: driver,
-			mockT:  testutil.NewMockT(t),
-			schema: schema,
+			dbPool:     dbPool,
+			driver:     driver,
+			mockT:      testutil.NewMockT(t),
+			schema:     schema,
+			schemaOpts: &RequireInsertedOpts{Schema: schema},
 		}
 	}
 
@@ -673,11 +679,75 @@ func TestRequireManyInserted(t *testing.T) {
 		_, err := riverClient.Insert(ctx, Job1Args{String: "foo"}, nil)
 		require.NoError(t, err)
 
-		jobs := requireManyInserted(ctx, bundle.mockT, bundle.driver, bundle.schema, []ExpectedJob{
-			{Args: &Job1Args{}},
+		jobs := requireManyInserted(ctx, bundle.mockT, bundle.driver, []ExpectedJob{
+			{Args: &Job1Args{}, Opts: bundle.schemaOpts},
 		})
 		require.False(t, bundle.mockT.Failed)
 		require.Equal(t, "job1", jobs[0].Kind)
+	})
+
+	t.Run("SchemaInAllOptsOkay", func(t *testing.T) {
+		t.Parallel()
+
+		riverClient, bundle := setup(t)
+
+		_, err := riverClient.InsertMany(ctx, []river.InsertManyParams{
+			{Args: Job1Args{String: "foo"}},
+			{Args: Job1Args{String: "bar"}},
+		})
+		require.NoError(t, err)
+
+		jobs := requireManyInserted(ctx, bundle.mockT, bundle.driver, []ExpectedJob{
+			{Args: &Job1Args{String: "foo"}, Opts: bundle.schemaOpts},
+			{Args: &Job1Args{String: "bar"}, Opts: bundle.schemaOpts},
+		})
+		require.False(t, bundle.mockT.Failed)
+		require.Equal(t, "job1", jobs[0].Kind)
+	})
+
+	t.Run("SchemaOnlyInFirstOptsPositionOkay", func(t *testing.T) {
+		t.Parallel()
+
+		riverClient, bundle := setup(t)
+
+		_, err := riverClient.InsertMany(ctx, []river.InsertManyParams{
+			{Args: Job1Args{String: "foo"}},
+			{Args: Job1Args{String: "bar"}},
+		})
+		require.NoError(t, err)
+
+		jobs := requireManyInserted(ctx, bundle.mockT, bundle.driver, []ExpectedJob{
+			{Args: &Job1Args{String: "foo"}, Opts: bundle.schemaOpts},
+			{Args: &Job1Args{String: "bar"}},
+		})
+		require.False(t, bundle.mockT.Failed)
+		require.Equal(t, "job1", jobs[0].Kind)
+	})
+
+	t.Run("SchemaBeyondFirstPositionError", func(t *testing.T) {
+		t.Parallel()
+
+		_, bundle := setup(t)
+
+		_ = requireManyInserted(ctx, bundle.mockT, bundle.driver, []ExpectedJob{
+			{Args: &Job1Args{}, Opts: &RequireInsertedOpts{}},
+			{Args: &Job2Args{}, Opts: &RequireInsertedOpts{Schema: bundle.schema}},
+		})
+		require.True(t, bundle.mockT.Failed)
+		require.Equal(t, bundle.mockT.LogOutput(), failureString("Internal failure: when setting RequireInsertedOpts.Schema with RequireMany schema should be set only at index 0 or the same schema set for all options; expectedJobs[0].Opts.Schema = %q, expectedJobs[%d].Opts.Schema = %q", "", 1, bundle.schema)+"\n")
+	})
+
+	t.Run("DifferingSchemaError", func(t *testing.T) {
+		t.Parallel()
+
+		_, bundle := setup(t)
+
+		_ = requireManyInserted(ctx, bundle.mockT, bundle.driver, []ExpectedJob{
+			{Args: &Job1Args{}, Opts: &RequireInsertedOpts{Schema: bundle.schema}},
+			{Args: &Job2Args{}, Opts: &RequireInsertedOpts{Schema: "other_schema"}},
+		})
+		require.True(t, bundle.mockT.Failed)
+		require.Equal(t, bundle.mockT.LogOutput(), failureString("Internal failure: when setting RequireInsertedOpts.Schema with RequireMany schema should be set only at index 0 or the same schema set for all options; expectedJobs[0].Opts.Schema = %q, expectedJobs[%d].Opts.Schema = %q", bundle.schema, 1, "other_schema")+"\n")
 	})
 }
 
@@ -711,7 +781,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		_, err := riverClient.InsertTx(ctx, bundle.tx, Job1Args{String: "foo"}, nil)
 		require.NoError(t, err)
 
-		jobs := requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, []ExpectedJob{
+		jobs := requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, []ExpectedJob{
 			{Args: &Job1Args{}},
 		})
 		require.False(t, bundle.mockT.Failed)
@@ -730,14 +800,14 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		require.NoError(t, err)
 
 		// Visible in the original transaction.
-		jobs := requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, []ExpectedJob{
+		jobs := requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, []ExpectedJob{
 			{Args: &Job1Args{}},
 		})
 		require.False(t, bundle.mockT.Failed)
 		require.Equal(t, "job1", jobs[0].Kind)
 
 		// Not visible in the second transaction.
-		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, otherTx, emptySchema, []ExpectedJob{
+		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, otherTx, []ExpectedJob{
 			{Args: &Job1Args{}},
 		})
 		require.True(t, bundle.mockT.Failed)
@@ -754,7 +824,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		_, err = riverClient.InsertTx(ctx, bundle.tx, Job2Args{Int: 123}, nil)
 		require.NoError(t, err)
 
-		jobs := requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, []ExpectedJob{
+		jobs := requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, []ExpectedJob{
 			{Args: &Job1Args{}},
 			{Args: &Job2Args{}},
 		})
@@ -774,7 +844,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		jobs := requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, []ExpectedJob{
+		jobs := requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, []ExpectedJob{
 			{Args: &Job1Args{}},
 			{Args: &Job1Args{}},
 		})
@@ -797,7 +867,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		jobs := requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, []ExpectedJob{
+		jobs := requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, []ExpectedJob{
 			{Args: &Job1Args{}},
 			{Args: &Job1Args{}},
 			{Args: &Job2Args{}},
@@ -821,7 +891,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		_, err := riverClient.InsertTx(ctx, bundle.tx, Job1Args{String: "foo"}, nil)
 		require.NoError(t, err)
 
-		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, []ExpectedJob{
+		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, []ExpectedJob{
 			{
 				Args: &Job1Args{},
 				Opts: &RequireInsertedOpts{
@@ -843,7 +913,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, []ExpectedJob{
+		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, []ExpectedJob{
 			{
 				Args: &Job2Args{},
 				Opts: &RequireInsertedOpts{
@@ -863,7 +933,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 
 		_, bundle := setup(t)
 
-		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, []ExpectedJob{
+		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, []ExpectedJob{
 			{Args: &Job1Args{}},
 		})
 		require.True(t, bundle.mockT.Failed)
@@ -883,7 +953,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, []ExpectedJob{
+		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, []ExpectedJob{
 			{Args: &Job1Args{}},
 		})
 		require.True(t, bundle.mockT.Failed)
@@ -903,7 +973,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, []ExpectedJob{
+		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, []ExpectedJob{
 			{Args: &Job1Args{}},
 			{Args: &Job2Args{}},
 		})
@@ -927,7 +997,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, emptySchema, []ExpectedJob{
+		_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, bundle.mockT, bundle.tx, []ExpectedJob{
 			{Args: &Job1Args{}},
 			{Args: &Job1Args{}},
 			{Args: &Job2Args{}},
@@ -958,7 +1028,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		// Max attempts
 		{
 			mockT := testutil.NewMockT(t)
-			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, []ExpectedJob{
+			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, []ExpectedJob{
 				{
 					Args: &Job2Args{},
 					Opts: &RequireInsertedOpts{
@@ -980,7 +1050,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		// Priority
 		{
 			mockT := testutil.NewMockT(t)
-			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, []ExpectedJob{
+			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, []ExpectedJob{
 				{
 					Args: &Job2Args{},
 					Opts: &RequireInsertedOpts{
@@ -1002,7 +1072,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		// Queue
 		{
 			mockT := testutil.NewMockT(t)
-			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, []ExpectedJob{
+			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, []ExpectedJob{
 				{
 					Args: &Job2Args{},
 					Opts: &RequireInsertedOpts{
@@ -1024,7 +1094,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		// Scheduled at
 		{
 			mockT := testutil.NewMockT(t)
-			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, []ExpectedJob{
+			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, []ExpectedJob{
 				{
 					Args: &Job2Args{},
 					Opts: &RequireInsertedOpts{
@@ -1046,7 +1116,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		// State
 		{
 			mockT := testutil.NewMockT(t)
-			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, []ExpectedJob{
+			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, []ExpectedJob{
 				{
 					Args: &Job2Args{},
 					Opts: &RequireInsertedOpts{
@@ -1068,7 +1138,7 @@ func TestRequireManyInsertedTx(t *testing.T) {
 		// Tags
 		{
 			mockT := testutil.NewMockT(t)
-			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, emptySchema, []ExpectedJob{
+			_ = requireManyInsertedTx[*riverpgxv5.Driver](ctx, mockT, bundle.tx, []ExpectedJob{
 				{
 					Args: &Job2Args{},
 					Opts: &RequireInsertedOpts{